// doctor/src/modules/doctor-certificates/enums/academic_degree.enum.ts
export enum AcademicDegree {
    PROFESSOR = 'PROFESSOR',            // GS
    ASSOCIATE_PROFESSOR = 'ASSOCIATE_PROFESSOR',  // PGS

    PHD = 'PHD',                        // TS
    MASTER = 'MASTER',                  // ThS
    BACHELOR = 'BACHELOR',              // CN

    MD = 'MD',                          // BS
    SPECIALIST_I = 'SPECIALIST_I',      // BSCKI
    SPECIALIST_II = 'SPECIALIST_II',    // BSCKII
}